+++ Config/config.php
@@ -39,7 +39,7 @@ $defaultIntegrationArguments = array_merge(
 return [
     'name'        => 'DOI Confirm Bundle',
     'description' => 'Adds a robust and flexible way to add a double-opt-in process (DOI) to any form in Mautic.',
-    'version'     => '3.0.0',
+    'version'     => '3.0.1',
     'author'      => 'Alexander Zlobin',
     'services' => [
         'events' => [